Small improvement

+++ scripts/stable_txt2img.py
@@ -550,7 +550,7 @@ def main(opt):
             if init_mask_path is not None:
                 mask_obj = Image.open(init_mask_path).convert("L")
                 print(f"Applied mask {init_mask_path}")
-                mask     = torch.from_numpy(np.array(mask_obj).astype(np.uint8)) / 255
+                mask     = torch.from_numpy(np.array(mask_obj)) / 255
                 mask     = F.interpolate(mask[None, None], size=(opt.H // opt.f, opt.W // opt.f), mode='nearest')
                 mask     = mask.repeat(batch_size, 1, 1, 1).to(device)
                 mask_dict = { 'fg_mask': mask, 'aug_mask': None }
